+++ CHANGELOG.md
@@ -23,7 +23,7 @@ and this project does NOT adhere to [Semantic Versioning](https://semver.org/spe
 * net.py: Add `insecure=False, no_proxy=False, timeout=3` to `get_public_ip()`
 * nodebb.py: Evaluate `no_proxy` in `get_data()`
 * rocket.py: add `insecure=False, no_proxy=False, timeout=3` to all functions
-
+* veeam.py: no more hardcoded `insecure` and also evaluate `no_proxy` in `get_token()`
 
 
 ## 2023112901